LPS-71559 Move store wrappers to its own module
